requirements.txt
astroid==2.2.5
certifi==2019.9.11
chardet==3.0.4
Django==2.2.5
django-extensions==2.2.1
djangorestframework==3.10.3
graphviz==0.13
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
pydotplus==2.0.2
pygraphviz==1.5
pylint==2.3.1
pyodbc==4.0.27
pyparsing==2.4.2
pytz==2019.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.3
wrapt==1.11.2
